backend/requirements.txt
fastapi>=0.134.0
uvicorn[standard]>=0.41.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.7.0
fastapi-mail>=1.4.0
apscheduler>=3.10.0
razorpay>=1.4.0
qrcode[pil]>=7.4.0
